refactor: file, folder and manage
